+++ R/updog.R
@@ -46,6 +46,8 @@
 #'   \item{\code{\link{multidog}}}{A convenience function for running
 #'       \code{\link{flexdog}} over many SNPs. This function provides
 #'       support for parallel computing.}
+#'   \item{\code{\link{format_multidog}}}{Return matricized elements from the output of \code{\link{multidog}}.}
+#'   \item{\code{\link{filter_snp}}}{Filter SNPs based on the output of \code{\link{multidog}}}
 #'   \item{\code{\link{mupdog}}}{An experimental approach to genotype
 #'       autopolyploids that accounts for varying levels of
 #'       relatedness between the individuals in the sample.}